Remove usages of BasicPillow in tests
