+++ corehq/ex-submodules/pillowtop/tests/__init__.py
@@ -4,6 +4,5 @@ from .test_checkpoints import *
 from .test_config import *
 from .test_couch import *
 from .test_dao import *
-from .test_es_reindex import *
 from .test_elasticsearch import *
 from .test_import_pillows import *
